Allow non-existent paths for new ini handling
Previously, validation of the requested ini was handled at the parameter level. However, some supporting functions I am adding will update an ini to specified values, creating the file if it doesn't exist.

In order to pipe the output from Get-IniContent to the supporting function, I needed Get-IniContent to handle the situation where the ini doesn't exist.

With this change, if the ini is not found, Get-IniContent simply returns early with the empty ordered dictionary instead of failing to process.