Get or Create the ini file in the Process block
When piping another ini editing function to Out-IniFile, the get or creation of the ini file was getting called *before* the ini updating was done since it was in the Begin block. So I was ending up with either blank ini's or ini's without the updated values. To fix this I moved the get or creation lines down to the Process block.