Skip entire newLine cleanup process of existing file
With this commit, there is no more reading the entire output file
and no checking of correctness of newLines for the entire file.

Also, this way, an empty line is inserted at the end of the file.

Requires consultation with project owner
and change in tests to allow for this empty line at the end of files