[Fix] Improve performance by operating in-memory (#80)
* [Fix] Improve performance by operating in-memory

This commit replaces all the Out-File operations,
with a list of strings to write at the end of programs (Export) execution.

All previous functionality, like "Encoding", "Appending", "Format", etc.
are handled.

* Further reduce amount of IO operations

* Reduce naming ambiguity & remove unused parameters

Removes no longer required fileParamters from Out-Keys collection script

* Example description specification

* [WIP] Don't add new line character at the end of file

In the original script, contents of the file were written into a file, then read back into program memory as a string.
Then additional NewLines were stripped from the file.

Because we are re-creating the entire structure from Objects array, there is no need to do this cleanup anymore.

* Skip entire newLine cleanup process of existing file

With this commit, there is no more reading the entire output file
and no checking of correctness of newLines for the entire file.

Also, this way, an empty line is inserted at the end of the file.

Requires consultation with project owner
and change in tests to allow for this empty line at the end of files

* Don't delete file on export if exists, just overwrite it

This stops updating "Created" & updates only "Modified" line in metadata

* Update tests for minified format & make tests accept empty new line at the end of pretty files

Additional empty line makes the appended data separated by an empty line.
BEAUTIFUL!

* Tag Remove-EmptyLine as obsolete