Added `-InputString` to `Import-Ini`
This allows a string representation of a ini file to be imported.
This can be useful when the file is read from a web request.

In order to implement this, a slight performance buff was implemented:
files are now read with `[System.IO.File]::ReadAllLines()` instead for `switch -file`
Read more: https://devdojo.com/hcritter/powershell-performance-test-file-reading

closes #62, closes #71